Move LinksTest to appropriate file
